summit

+++ _strlen.c
@@ -4,9 +4,9 @@
 * @str: string
 * Return: size of string
 */
-size_t _strlen(char *str)
+int _strlen(char *str)
 {
-	size_t i;
+	int i;
 
 	for (i = 0; str[i]; i++)
 	{
